Move Grid.java to engine package and set Plant coordinates before attacking
